requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
Django==3.1.4
django-autoslug==1.9.8
django-crispy-forms==1.11.2
idna==2.10
paypal==1.2.5
Pillow==8.2.0
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
